Make rev_parse take care of to-string conversion
So far our GitRepo::rev_parse method was just a thin wrapper around
git_out and just forwarded the "raw" output data. That caused a bunch of
duplication of to string conversions at client sites.
With this change we adjust the method to take care of the to string
conversion itself instead.